lexical_wizard_cs410_project/requirements.txt
langchain
openai
PyPDF2
faiss-cpu
huggingface-hub
python-dotenv
sentence-transformers
streamlit
replicate
